+++ spec/features/early_career_payments_claim_spec.rb
@@ -282,6 +282,10 @@ RSpec.feature "Teacher Early-Career Payments claims" do
     expect(claim.reload.student_loan_start_date).to eq(StudentLoan::BEFORE_1_SEPT_2012)
     expect(claim.student_loan_plan).to eq(StudentLoan::PLAN_1)
 
+    # - Are you currently paying off your masters/doctoral loan
+    expect(page).not_to have_text(I18n.t("questions.has_masters_and_or_doctoral_loan"))
+    expect(claim.reload.has_masters_doctoral_loan).to be_nil
+
     # - Did you take out a postgraduate masters loan on or after 1 August 2016
     expect(page).to have_text(I18n.t("questions.postgraduate_masters_loan"))
 
@@ -732,6 +736,10 @@ RSpec.feature "Teacher Early-Career Payments claims" do
     expect(claim.reload.student_loan_start_date).to eq(StudentLoan::BEFORE_1_SEPT_2012)
     expect(claim.student_loan_plan).to eq(StudentLoan::PLAN_1)
 
+    # - Are you currently paying off your masters/doctoral loan
+    expect(page).not_to have_text(I18n.t("questions.has_masters_and_or_doctoral_loan"))
+    expect(claim.reload.has_masters_doctoral_loan).to be_nil
+
     # - Did you take out a postgraduate masters loan on or after 1 August 2016
     expect(page).to have_text(I18n.t("questions.postgraduate_masters_loan"))
 
@@ -1102,6 +1110,10 @@ RSpec.feature "Teacher Early-Career Payments claims" do
 
       expect(claim.reload.student_loan_start_date).to eq(StudentLoan::BEFORE_1_SEPT_2012)
       expect(claim.student_loan_plan).to eq(StudentLoan::PLAN_1)
+
+      # - Are you currently paying off your masters/doctoral loan
+      expect(page).not_to have_text(I18n.t("questions.has_masters_and_or_doctoral_loan"))
+      expect(claim.reload.has_masters_doctoral_loan).to be_nil
 
       # - Did you take out a postgraduate masters loan on or after 1 August 2016
       expect(page).to have_text(I18n.t("questions.postgraduate_masters_loan"))