+++ CHANGELOG.md
@@ -46,6 +46,11 @@ The format is based on [Keep a Changelog]
   - 'What gender does your school's payroll system associate with you?' content
     change
   - Update subject heading on TSLR and ECP OTP email
+  - Split Postgraduate loans (ITT course) from Postgraduate Masters/Doctoral
+    loans
+  - Update Postgraduate Masters and Doctoral Loans questions to be only being
+    asked if the answer to 'Do you have postgraduate masters and/or doctoral'
+    loans is 'Yes'
 - ECP policies
   - Adds set reminder flow to eligible now flow
   - Fix no validation occuring on 'Personal Details' screens